rework api test cases
