+++ tests/api/__init__.py
@@ -1,0 +1,8 @@
+"""Defines stable test cases for the public API.
+
+These are basically integration tests, in that they don't isolate their
+dependencies, and just execute tests directly on the public exposed functions
+of this library.
+
+These tests cases should change only with respect to semantic versioning.
+"""
